--- sme_ofertaimoveis/users/api/login.py
@@ -1,60 +1,0 @@
-import logging
-import json
-
-from django.contrib.auth import get_user_model
-from rest_framework import permissions, status
-from rest_framework.response import Response
-from rest_framework_jwt.views import ObtainJSONWebToken
-
-from ...users.api.services import AutenticacaoService
-
-User = get_user_model()
-logger = logging.getLogger(__name__)
-
-
-class LoginView(ObtainJSONWebToken):
-    """
-    POST auth/login/
-    """
-
-    permission_classes = (permissions.AllowAny,)
-
-    def post(self, request, *args, **kwargs):
-        login = request.data.get("login", "")
-        senha = request.data.get("senha", "")
-
-        try:
-            response = AutenticacaoService.autentica(login, senha)
-            if response.status_code == 200:
-                user_dict = response.json()
-                if 'login' in user_dict.keys():
-                    try:
-                        user = User.objects.get(username=user_dict['login'])
-                        user.name = user_dict['nome']
-                        user.email = user_dict['email']
-                        user.set_password(senha)
-                        user.save()
-                    except User.DoesNotExist:
-                        # logger.info("Usuário %s não encontrado.", login)
-                        # return Response({'data': {'detail': 'Usuário não encontrado.'}}, status=status.HTTP_401_UNAUTHORIZED)
-                        user = User()
-                        user.name = user_dict['nome']
-                        user.email = user_dict['email']
-                        user.set_password(senha)
-                        user.save()
-                    request._full_data = {'username': user_dict['login'], 'password': senha}
-                    resp = super().post(request, *args, **kwargs)
-                    user_dict['permissoes'] = self.get_user_permissions(user)
-                    data = {**user_dict, **resp.data}
-                    return Response(data)
-            return Response(response.json(), response.status_code)
-        except Exception as e:
-            return Response({'data': {'detail': f'ERROR - {e}'}}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    def get_user_permissions(self, user):
-        perms = []
-        for group in user.groups.all():
-            for permission in group.permissions.all():
-                perms.append(permission.codename)
-
-        return perms
