+++ sme_ofertaimoveis/users/api/serializers.py
@@ -11,4 +11,4 @@ class UserSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = User
-        fields = ("username", "email",)
+        fields = ("username", "email", "perfil_usuario")
