+++ sme_ofertaimoveis/users/api/viewsets.py
@@ -7,8 +7,10 @@ from rest_framework.mixins import RetrieveModelMixin, ListModelMixin, UpdateMode
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
+from rest_framework_jwt.views import ObtainJSONWebToken
 
 from .serializers import UserSerializer
+from .services import AutenticacaoService
 
 User = get_user_model()
 
@@ -89,3 +91,51 @@ class UsuarioConfirmaEmailViewSet(GenericViewSet):
             return Response({'detail': 'Parâmetros para confirmação de e-mail inválidos'},
                             status=status.HTTP_400_BAD_REQUEST)
         return Response(UserSerializer(usuario).data)
+
+
+class LoginView(ObtainJSONWebToken):
+    """
+    POST auth/login/
+    """
+
+    permission_classes = (permissions.AllowAny,)
+
+    def post(self, request, *args, **kwargs):
+        login = request.data.get("login", "")
+        senha = request.data.get("senha", "")
+        try:
+            response = AutenticacaoService.autentica(login, senha)
+            if response.status_code == 200:
+                user_dict = response.json()
+                if 'usuarioId' in user_dict.keys():
+                    try:
+                        user = User.objects.get(username=login)
+                        user.first_name = user_dict['nome'].split(' ')[0]
+                        user.last_name = ' '.join(user_dict['nome'].split(' ')[1:])
+                        user.set_password(senha)
+                        user.save()
+                    except User.DoesNotExist:
+                        # logger.info("Usuário %s não encontrado.", login)
+                        # return Response({'data': {'detail': 'Usuário não encontrado.'}}, status=status.HTTP_401_UNAUTHORIZED)
+                        user = User()
+                        user.username = login
+                        user.first_name = user_dict['nome'].split(' ')[0]
+                        user.last_name = ' '.join(user_dict['nome'].split(' ')[1:])
+                        user.set_password(senha)
+                        user.save()
+                    request._full_data = {'username': login, 'password': senha}
+                    resp = super().post(request, *args, **kwargs)
+                    user_dict['permissoes'] = self.get_user_permissions(user)
+                    data = {**user_dict, **resp.data}
+                    return Response(data)
+            return Response(response.json(), response.status_code)
+        except Exception as e:
+            return Response({'data': {'detail': f'ERROR - {e}'}}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+    def get_user_permissions(self, user):
+        perms = []
+        for group in user.groups.all():
+            for permission in group.permissions.all():
+                perms.append(permission.codename)
+
+        return perms
